backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Geospatial
shapely==2.0.2
fastkml==1.4.0
lxml==4.9.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0  # Pin to 3.2.0 for passlib compatibility
python-dotenv==1.0.0

# HTTP requests
httpx==0.25.2

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Rate limiting (optional)
# slowapi==0.1.9
